full working pipeline from extract to training and saving deep learning model; currently the one in checkpoint is only trained over 100 epochs; run main.py to run the full pipeline
