+++ ETL/__init__.py
@@ -1,3 +1,4 @@
 from .processData import DataProcessor
 from .extractData import GazeDataExtractor
 from .transformData import GazeDataTransformer
+from .userGazeVideoCreator import UserGazeVideoCreator
