+++ models/.gitignore
@@ -1,2 +1,3 @@
 models/*.pkl
 *.pkl
+checkpoints/*
