+++ models/__init__.py
@@ -1,2 +1,3 @@
 from .naiveModel import NaiveHazardDetector
+from .deep_learning_train import DeepLearningModel
