+++ requirements.txt
@@ -7,5 +7,13 @@ scipy==1.11.4
 opencv-python==4.11.0.86
 numpy==1.23.5
 matplotlib==3.7.1
-scikit-learn==1.2.2
+scikit-learn==1.6.1
+statsmodels==0.14.0
+transformers==4.31.0
+torch==2.6.0
+torchvision==0.21.0
+pillow==10.4.0
+tqdm==4.67.1
+opencv-python==4.11.0.86
+certifi
 ipykernel
